Update Final PS Script.txt
This update ensures the model backup process is based on the Workspace type:

Premium and Fabric Workspaces: The model is downloaded directly via the workspace connection using Tabular Editor's CLI.

Pro Workspaces: The script downloads the report PBIX via the Get-PowerBIReport command and then leverages pbi-tools to extract the model into a folder structure and build a .bim file.

Previously, the script would check if there was already a bim file and use the Pro workspace method if no bim file was found. This caused unnecessary extractions on Premium/Fabric capacity 'thin' reports that are only connected to the model.

This now checks the Workspace attribute and only uses the pbi-tools method if the Workspace is not on a dedicated capacity. If dedicatedCapacity is false, it indicates a Pro workspace, and the pbi-tools method is used.