Adds 'My Workspace' to solution
Adds 'My Workspace' to solution

Also identifies if a report is shared to the user but they do not have Workspace access.

Adds a check to ensure no report is downloaded twice if it's shared to them via an app but they already have workspace access. The REST API puts any Report shared to a user and shared via an organizational app under the My Workspace list of reports. This correctly filters those out or identifies them as 'Shared Reports - No Workspace Access'